requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.0.8
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
mixins==0.1.4
mysqlclient @ file:///C:/Users/user/Ceos/django-rest-framework-13th/mysqlclient-1.4.6-cp36-cp36m-win32.whl
nose==1.3.7
Pillow==8.1.2
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
